// src/services/wallet-state-core/generated/BhistShellResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BhistEntry } from "./BhistEntry";

/**
 * What the shell observed.
 */
export type BhistShellResult = { "type": "loaded", entries: Array<BhistEntry>, } | { "type": "written" };